+++ Sources/Sentry/SentrySDKInternal.m
@@ -726,7 +726,8 @@ static NSDate *_Nullable startTimestamp = nil;
             return;
         }
 
-        if (sentry_profileConfiguration.profilerSessionSampleDecision.decision
+        SentryProfileConfiguration *profileConfiguration = sentry_getProfileConfiguration();
+        if (profileConfiguration.profilerSessionSampleDecision.decision
             != kSentrySampleDecisionYes) {
             SENTRY_LOG_DEBUG(
                 @"The profiling session has been sampled out, no profiling will take place.");
@@ -746,14 +747,15 @@ static NSDate *_Nullable startTimestamp = nil;
 {
     // check if we'd be stopping a launch profiler, because then we need to check the hydrated
     // configuration options, not the current ones
-    if (sentry_profileConfiguration.isProfilingThisLaunch) {
-        if (sentry_profileConfiguration.isContinuousV1) {
+    SentryProfileConfiguration *profileConfiguration = sentry_getProfileConfiguration();
+    if (profileConfiguration.isProfilingThisLaunch) {
+        if (profileConfiguration.isContinuousV1) {
             SENTRY_LOG_DEBUG(@"Stopping continuous v1 launch profile.");
             [SentryContinuousProfiler stop];
             return;
         }
 
-        if (sentry_profileConfiguration.profileOptions == nil) {
+        if (profileConfiguration.profileOptions == nil) {
             SENTRY_LOG_WARN(
                 @"The current profiler was started on app launch and was configured as a "
                 @"transaction profiler, which cannot be stopped manually. Transaction profiling is "
@@ -761,7 +763,7 @@ static NSDate *_Nullable startTimestamp = nil;
             return;
         }
 
-        if (sentry_profileConfiguration.profileOptions.lifecycle == SentryProfileLifecycleTrace) {
+        if (profileConfiguration.profileOptions.lifecycle == SentryProfileLifecycleTrace) {
             SENTRY_LOG_WARN(
                 @"The launch profile lifecycle was set to trace, so you cannot stop profile "
                 @"sessions manually. See SentryProfileLifecycle for more information.");